Combine some query templates
